pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nlq-validator"
version = "0.1.0"
description = "Natural Language Query validator"
requires-python = ">=3.10"
dependencies = [
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
dev       = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]
openai    = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
gemini    = ["google-generativeai>=0.8"]
mistral   = ["mistralai>=1.0"]
all-llm    = ["openai>=1.0", "anthropic>=0.25", "google-generativeai>=0.8", "mistralai>=1.0"]
embeddings = ["sentence-transformers>=2.7"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
